requirement.txt
Package     Version
----------- -------
numpy       1.19.5
Pillow      8.1.0
pip         21.0.1
setuptools  41.2.0
